requirements.txt
discord.py
aiomcrcon
Babel